Reorganize and change properties of FancyConsole
