+++ src/io/github/nateyoung/frederic/Frederic.java
@@ -1,4 +1,4 @@
-package src;
+package io.github.nateyoung.frederic;
 
 import java.io.File;
 import java.io.IOException;
@@ -8,6 +8,13 @@ import java.util.ArrayList;
 import java.util.List;
 
 import javax.swing.JFrame;
+
+import io.github.nateyoung.frederic.command.DoCommand;
+import io.github.nateyoung.frederic.command.DoWithCommand;
+import io.github.nateyoung.frederic.console.FCRead;
+import io.github.nateyoung.frederic.console.FancyConsole;
+import io.github.nateyoung.frederic.util.WeightedRandomizer;
+
 /**This class is the central class for the Frederic program. <br>
  * Client programs should usually construct only one instance of this class. Each <code>Frederic</code> 
  * instance runs separately from all other <code>Frederic</code> instances and has its own FancyConsole of
