+++ src/io/github/nateyoung/frederic/command/DoWithCommand.java
@@ -1,6 +1,8 @@
-package src;
+package io.github.nateyoung.frederic.command;
 
 import java.io.File;
+
+import io.github.nateyoung.frederic.Frederic;
 /**
  * A simple interface to represent a possible action to be taken when the user enters "do something with" 
  * and a file name.
